+++ packages/docs/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @finsweet/attributes-docs
 
+## 1.7.13
+
+### Patch Changes
+
+- 7d626e05: Enabled support for socialshare and queryparams
+
+## 1.7.12
+
+### Patch Changes
+
+- Updated dependencies [7542dcdb]
+  - @global/constants@1.0.2
+  - @global/factory@1.1.3
+  - @finsweet/attributes-codehighlight@1.3.5
+  - @finsweet/attributes-copyclip@1.5.5
+  - @finsweet/attributes-toc@1.0.11
+
+## 1.7.11
+
+### Patch Changes
+
+- Updated dependencies [bdd1a78]
+  - @global/constants@1.0.1
+  - @global/factory@1.1.2
+  - @finsweet/attributes-codehighlight@1.3.4
+  - @finsweet/attributes-copyclip@1.5.4
+  - @finsweet/attributes-toc@1.0.10
+
 ## 1.7.10
 
 ### Patch Changes